refactor(be): move get_ids() method to utils.py
