+++ backend/core/utils.py
@@ -1,6 +1,8 @@
 # utils.py
 import os
 import requests
+from dotenv import load_dotenv
+
 
 def compress_image_tinify(image_file):
     """
@@ -30,3 +32,29 @@ def compress_image_tinify(image_file):
     except Exception as e:
         print(f"Error compressing image: {str(e)}")
         return None
+    
+
+load_dotenv()
+api_key = os.getenv('BABELNET_API_KEY')
+
+
+def get_ids(word_id):
+    return_array = [word_id]
+
+    url = 'https://babelnet.io/v9/getOutgoingEdges'
+    params = {
+        'id': word_id,
+        'key': api_key,
+    }
+    response = requests.get(url, params=params)
+    if response.status_code != 200:
+        raise Exception(f"Error fetching from BabelNet API. Status code: {response.status_code}")
+    
+    data = response.json()
+    for value in data:
+        if value.get("language") == "EN" or value.get("language") == "TR":
+            # if value.get("pointer").get("shortName") != "related":
+            #     return_array.append(value.get("target"))
+            return_array.append(value.get("target"))    
+    
+    return return_array
